Arrange samlassertion package
